requirements.txt
# API framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Data processing and ML
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.0
scipy==1.11.2

# Testing
pytest==7.4.2

# Deployment
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional for TF-IDF implementation
nltk==3.8.1